+++ recipes/default.rb
@@ -1,8 +1,23 @@
-#
-# Cookbook Name:: rbenv
-# Recipe:: default
-#
-# Copyright (C) 2015 YOUR_NAME
-#
-# All rights reserved - Do Not Redistribute
-#
+include_recipe 'apt'
+include_recipe 'build-essential'
+include_recipe 'git'
+
+package 'zlib1g'
+package 'zlib1g-dev'
+
+package 'openssl'
+package 'libssl-dev'
+
+package 'libc6-dev'
+package 'automake'
+package 'libtool'
+package 'libyaml-dev'
+
+package 'libreadline-dev'
+
+package 'libxml2-dev'
+package 'libxslt1-dev'
+
+package 'ncurses-dev'
+
+package 'libffi-dev'