+++ resources/source.rb
@@ -1,0 +1,3 @@
+actions :install
+default_action :install
+attribute :user, :kind_of => String, :required => true, :name_attribute => true